refactor scripts for better testability
