+++ action-scripts/build-config.js
@@ -1,32 +1,14 @@
 const fs = require("fs");
-const path = require("path");
 
-const args = process.argv.slice(2);
-
-const copyThemeFile = async ({ dist, styleSystem, actionPath }) => {
-  const themeTemplate = {
-    tailwind: "tailwind.config.js",
-    mui: "muiTheme.js",
-  }[styleSystem];
-
-  const themeSource = path.join(actionPath, "theme_templates", themeTemplate);
-  const themeDest = path.join(dist, styleSystem, themeTemplate);
-
-  await fs.copyFile(themeSource, themeDest, (err) =>
-    console.warn(`Copy File Error: ${err}`)
-  );
-};
-
-const buildConfig = async (
-  dist = "build",
-  src = "tokens",
-  styleSystem = "tailwind",
-  actionPath = "" // TODO: Use actionPath to access theme objects and copy to build directory
-) => {
+const buildConfig = async ({
+  outputFolder = "build",
+  transformedTokenPath,
+  styleSystem,
+}) => {
   const platforms = {
     tailwind: {
       transformGroup: "js/custom",
-      buildPath: `${dist}/`,
+      buildPath: `${outputFolder}/`,
       files: [
         {
           destination: "typography.json",
@@ -48,7 +30,7 @@ const buildConfig = async (
     },
     mui: {
       transformGroup: "js/custom",
-      buildPath: `${dist}/`,
+      buildPath: `${outputFolder}/`,
       files: [
         {
           destination: "typography.json",
@@ -71,7 +53,7 @@ const buildConfig = async (
   };
 
   const config = {
-    source: [`${src}/*.json`],
+    source: [transformedTokenPath],
     platforms: { [styleSystem]: platforms[styleSystem] },
   };
 
@@ -80,8 +62,6 @@ const buildConfig = async (
       console.error(err);
     }
   });
-
-  await copyThemeFile({ dist, styleSystem, actionPath });
 };
 
-buildConfig(args[0], args[1], args[2], args[3]);
+module.exports = { buildConfig };
