+++ action.yml
@@ -26,14 +26,16 @@ runs:
         ref: ${{ github.head_ref }}
     - name: Setup Node.js environment
       uses: actions/setup-node@v2.4.0
+
+      # install prod deps
     - run: npm install --omit=dev --prefix ${{ github.action_path }}
       shell: bash
-    - run: node ${{ github.action_path }}/build-config.js ${{ inputs.outputFolder }} ${{ inputs.outputFolder }}/style-dictionary ${{inputs.styleSystem}} ${{github.action_path}}
+
+      # run main.js script
+    - run: node ${{ github.action_path }}/main.js ${{ inputs.tokenFile }} ${{ inputs.outputFolder }} ${{inputs.styleSystem}} ${{github.action_path}}
       shell: bash
-    - run: npx token-transformer ${{ inputs.tokenFile }} ${{ inputs.outputFolder }}/style-dictionary/tokens.json
-      shell: bash
-    - run: node ${{ github.action_path }}/build.js
-      shell: bash
+
+      # commit new files
     - uses: stefanzweifel/git-auto-commit-action@v4
       with:
         commit_message: Updated style Shipwright tokens
